+++ CHANGELOG.md
@@ -1,17 +1,33 @@
-## 0.0.1 (March 23, 2011)
+## 0.5.1 (April 27, 2011)
 
 Features:
 
-  - First commit
-  
-## 0.3.20 (April 26, 2011)
+  - Fixed compound intervals quality error
+
+
+## 0.5.0 (April 26, 2011)
 
 Features:
 
-  - Added support to full diatonic scale in quality method
-  
+  - Added short notation method
+
+
 ## 0.4.0 (April 26, 2011)
 
 Features:
 
   - Added support for notes alterations
+
+
+## 0.3.20 (April 26, 2011)
+
+Features:
+
+  - Added support to full diatonic scale in quality method
+
+
+## 0.0.1 (March 23, 2011)
+
+Features:
+
+  - First commit